feat: add opentofu-fmt formatter.
This adds the `opentofu-fmt` formatter, in addition to terraform. It is
mutually exclusive with the terraform one, as both are use to format the
terraform filetypes.

Most of the code is duplicate from the terraform one, but I don't think
there is a simpler/more straightforward way to include opentofu in the
codebase.

Opentofu is a fork which maintains the open source Apache 2.0 license.
See its [manifesto](https://opentofu.org/manifesto/)

Not sure if relevant, but I would be happy to maintain
this formatter.