fix(erb_lint): bug where buffer is clobbered by erb_lint formatter
An output:match is performed to detect the header of the linter output
to exclude it from being injected into the buffer. However, it matches
on `==`, which is also a legitimate Ruby operator.
Because the linter output uses 16 `=` symbols in its header, adding an
additional `==` to the match will still work while not causing false
matches with legitimate uses of `==`.