fix: detect if `supports_method` is being called as a method or a function
Neovim is moving towards dropping the old function style (`.`) of
invoking these in favor of method style (`:`). We added support for that
in
https://github.com/nvimtools/none-ls.nvim/commit/94d024e6550e53c70c3ab765c476a0158b65f564,
but it's not perfect: it incorrectly assumes that if you're using
nvim-0.11 then *all* invocations of these helpers are method style.
I imagine that's true for core-neovim, but there are still plugins in
the ecosystem using `.` style. For example, see
<https://github.com/lukas-reineke/lsp-format.nvim/blob/v2.7.1/lua/lsp-format/init.lua#L260>,
which I use with Neovim nightly.